feat(explain-aggregation): support for sharded find in planner
