+++ packages/mongodb-explain-compat/index.d.ts
@@ -2,6 +2,6 @@ declare function convertExplain<T>(input: T): T;
 export = convertExplain;
 
 export function getStageCursorKey(stage: any): string | undefined;
-
 export function isAggregationExplain(input: any): boolean;
 export function isShardedAggregationExplain(input: any): boolean;
+export function isShardedFindExplain(input: any): boolean;
